SaloonShoot iOS game using Cocos2d-v3
Code for the iOS game SaloonShoot made using cocos2d-v3.
Uses Admob with mediation to show iAd and Admob ads.

NOTE: I am not working on this game anymore. I am uploading it just in
case it is of any help to anybody. So I may not be able to fix
bugs/issues, if any or give any kind of support.